despawn enemies with lifetime
