+++ src/data/projectiles.rs
@@ -17,9 +17,6 @@ pub enum AttackEffect {
     Status(StatusEffect),
 }
 
-#[derive(Component, Clone, Debug, Reflect, PartialEq, Eq)]
-pub struct Lifetime(pub Timer);
-
 #[derive(Component, Copy, Clone, Debug, Reflect, PartialEq, Eq)]
 pub struct Droplet(pub LiquidType);
 
@@ -51,11 +48,5 @@ impl DamageType {
             DamageType::Lightning => StatusEffect::Electrified,
             DamageType::Chemical => StatusEffect::Acidic,
         }
-    }
-}
-
-impl Lifetime {
-    pub fn new(duration: f32) -> Self {
-        Lifetime(Timer::from_seconds(duration, bevy::time::TimerMode::Once))
     }
 }
